Made StructType extend Seq[StructField], and moved SQLUserDefinedType into sql.types package.
